Remove redundant files
These two were created when I was thinking about extending the Date
and Duration objects from the time library.

I may well still do this, not being able to work with durations of
a month and over is a particular pain point and would like to be
able add/sub durations in this range from a Date object. Equally
determining a duration span in this month+ range between two dates
is of value