Date Arithmetic first pass
This isn't totally working yet but it's been painful so want to
stage gate where I'm up to.

Have now implemented the idea that date aritmetic can result in
multiple equivalent outputs, via the DateaArithmeticOutput object.
This is simply a thin wrapper over a vector of dates. The duration
object now returns this object, rather than a Date directly, to
accommodate the possibility of haveing multiple answers.

The multiple answers aren't returned as yet but I have been able to
write a method to determine the dates and durations combinations
that will multiply. Should just be a case of finding the alternate
answers where this function is true.