Change Periodicity
First implementation of change periodicity. This allows you to
convert a timeseries that has one periodicity (e.g. monthly)
to the same date range with a different periodicity (e.g. annual)

There a potentially a number of different methods for how to do
this but have only implemented additition so far. When going up
(small time periods to bigger time periods) this will add together
all the source periods that relate to each target period (e.g. the
12 months of the  year).

When going down it is potentially more tricky, as you have to infer
detail that is not there. Have assumed that target periods have
constant value (i.e. we can divide down the source value to calculate
the target values for each period)

I had to limit it to underlying type of f64 as this was the only type
that I could divide though. Will need to give som thought to whether
other types could be supported (esp. integers)

Need to implement other aggregation methods in future commits