Aggregate up by splitting staddling periods
If a source period straddles a pair fo target periods, the
aggregation will now split the source period proporionate to the
day count as applies to the two parts.

This only affects weeks to months/quarters/years as every other
type of transformation fully fits source periods into target
periods. This change only applied when aggregating upwards, and a
similar change needs to be applied when changing down